+++ libs/spartan/sonner/src/lib/hlm-toaster.ts
@@ -114,7 +114,7 @@ export class HlmToaster {
       ...options,
       classes: {
         ...options?.classes,
-        toast: hlm('rounded-2xl!', options?.classes?.toast),
+        toast: hlm('rounded-xl!', options?.classes?.toast),
       },
     };
   });